+++ Makefile.am
@@ -1197,6 +1197,8 @@ imap_lmtpd_SOURCES = \
 
 nodist_imap_lmtpd_SOURCES = imap/lmtp_err.c
 
+imap_lmtpd_CFLAGS = -DBUILD_LMTPD $(CFLAGS)
+
 if AUTOCREATE
 imap_lmtpd_SOURCES += \
 	imap/autocreate.c \
@@ -1209,7 +1211,8 @@ imap_lmtpd_SOURCES += imap/lmtp_sieve.c imap/lmtp_sieve.h imap/smtpclient.c
 if JMAP
 imap_lmtpd_SOURCES += \
 	imap/jmap_util.c imap/jmap_mail_query.c imap/jmap_mail_query_parse.c \
-	imap/caldav_util.c imap/dav_util.c imap/itip_support.c imap/zoneinfo_db.c
+	imap/caldav_util.c imap/dav_util.c imap/itip_support.c imap/zoneinfo_db.c \
+	imap/jmap_notif.c imap/jmap_ical.c
 endif # JMAP
 
 endif # SIEVE
@@ -1340,6 +1343,8 @@ imap_httpd_SOURCES += \
 	imap/jmap_mailbox.c \
 	imap/jmap_mdn.c \
 	imap/jmap_notes.c \
+	imap/jmap_notif.c \
+	imap/jmap_notif.h \
 	imap/jmap_push.c \
 	imap/jmap_push.h \
 	imap/jmap_util.c \
