+++ imap/http_dav.h
@@ -66,20 +66,6 @@
 #define SCHED_DEFAULT   "Default/"
 #define MANAGED_ATTACH  "Attachments/"
 
-/* XML namespace URIs */
-#define XML_NS_DAV      "DAV:"
-#define XML_NS_CALDAV   "urn:ietf:params:xml:ns:caldav"
-#define XML_NS_CARDDAV  "urn:ietf:params:xml:ns:carddav"
-#define XML_NS_ISCHED   "urn:ietf:params:xml:ns:ischedule"
-#define XML_NS_CS       "http://calendarserver.org/ns/"
-#define XML_NS_MECOM    "http://me.com/_namespace/"
-#define XML_NS_MOBME    "urn:mobileme:davservices"
-#define XML_NS_APPLE    "http://apple.com/ns/ical/"
-#define XML_NS_USERFLAG "http://cyrusimap.org/ns/userflag/"
-#define XML_NS_SYSFLAG  "http://cyrusimap.org/ns/sysflag/"
-#define XML_NS_DAVMOUNT "http://purl.org/NET/webdav/mount/"
-#define XML_NS_JMAPCAL  "urn:ietf:params:jmap:calendars"
-
 #define USER_COLLECTION_PREFIX  "user"
 #define GROUP_COLLECTION_PREFIX "group"
